requirements.txt
hexdump==3.3
lief==0.14.1
pycryptodome==3.20.0
unicorn==2.0.1.post1
